Synchronize gaze to known video timing
This is to account for a "feature" of PTB where it flips whatever frame it thinks ought to be on screen at this time, without reporting the actual frame number flipped. So we're assuming that the frame shown is whatever frame was "supposed" to be on screen at the time of flip.