Align stimName standards for TC and MW
There was this bug where TC data failed to strip out the file extension early in the pipeline, so I had to add it back in in several locations, while MW had the file extension stripped out - due to a case-sensitive check for '.MOV' that ignored '.mov'. Now everything ignores the file extension.